idkidkidk:(

+++ htab_size.c
@@ -6,6 +6,6 @@
 #include "htab_private.h"
 
 // Vrati pocet zaznamov v tabulke
-unsigned int htab_size(const htab_t * t) {
+size_t htab_size(const htab_t * t) {
     return t->size;
 }
